+++ src/azure-cli/azure/cli/command_modules/security/_params.py
@@ -5,7 +5,8 @@
 
 # pylint: disable=line-too-long
 
-from azure.cli.core.commands.parameters import resource_group_name_type
+from azure.cli.core.commands.parameters import (get_three_state_flag,
+                                                resource_group_name_type)
 from knack.arguments import CLIArgumentType
 from ._validators import (validate_alert_status,
                           validate_auto_provisioning_toggle,
@@ -17,6 +18,9 @@ location_arg_type = CLIArgumentType(options_list=('--location', '-l'), metavar='
 
 # Alerts
 alert_status_arg_type = CLIArgumentType(options_list=('--status'), metavar='STATUS', help='target status of the alert. possible values are "dismiss" and "activate"')
+
+# Atp
+storage_account_arg_type = CLIArgumentType(options_list=('--storage-account'), metavar='NAME', help='Name of an existing storage account.')
 
 # Auto Provisioning
 auto_provisioning_auto_provision_arg_type = CLIArgumentType(options_list=('--auto-provision'), metavar='AUTOPROVISION', help='Automatic provisioning toggle. possible values are "On" or "Off"')
@@ -36,6 +40,7 @@ workspace_setting_target_workspace_arg_type = CLIArgumentType(options_list=('--t
 
 def load_arguments(self, _):
     for scope in ['alert',
+                  'atp',
                   'task',
                   'setting',
                   'contact',
@@ -58,6 +63,9 @@ def load_arguments(self, _):
             c.argument(
                 'location',
                 arg_type=location_arg_type)
+            c.argument(
+                'storage_account_name',
+                arg_type=storage_account_arg_type)
 
     for scope in ['alert update']:
         with self.argument_context('security {}'.format(scope)) as c:
@@ -72,6 +80,10 @@ def load_arguments(self, _):
                 'auto_provision',
                 validator=validate_auto_provisioning_toggle,
                 arg_type=auto_provisioning_auto_provision_arg_type)
+
+    for scope in ['atp storage update']:
+        with self.argument_context('security {}'.format(scope)) as c:
+            c.argument('is_enabled', help='Enable or disable Advanced Threat Protection for a received storage account.', arg_type=get_three_state_flag())
 
     for scope in ['contact create']:
         with self.argument_context('security {}'.format(scope)) as c: