+++ src/azure-cli/HISTORY.rst
@@ -30,6 +30,12 @@ Release History
 
 * Add `monitor` subgroup to manage Log Analytics monitoring in Azure Red Hat OpensShift cluster
 
+**CDN**
+
+* Add support for rulesEngine feature
+* Add new commands group 'cdn endpoint rule' to manage rules
+* Update azure-mgmt-cdn version to 4.0.0 to use api version 2019-04-15
+
 **IoT**
 
 * Deprecated 'IoT hub Job' commands.
@@ -46,7 +52,12 @@ Release History
 
 * Fix #2092: az network dns record-set add/remove: add warning when record-set is not found. In the future, an extra argument will be supported to confirm this auto creation.
 
+**Security**
+
+* Added new commands `az atp show` and `az atp update` to view and manage advanced threat protection settings for storage accounts.
+
 **SQL**
+
 * [BREAKING CHANGE] `az sql db create`: Remove "WideWorldImportersStd" and "WideWorldImportersFull" as documented allowed values for "az sql db create --sample-name". These sample databases would always cause creation to fail.
 
 **Storage**